fix test issues

+++ src/ch8_sem1_deadlock.c
@@ -85,7 +85,7 @@ int main()
 	}
 	int threads[THREAD_N] = {};
 	for (int i = 0; i < THREAD_N; i++) {
-		threads[i] = thread_create(deadlock_test, (void *)i);
+		threads[i] = thread_create(deadlock_test, (void *)(long) i);
 	}
 	sleep(500);
 	for (int i = 0; i < THREAD_N; i++) {
